refactor: userEditForm
